SearchBar renamed
